requirements.txt
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-bootstrap-templates==1.2.4
plotly==5.24.1
pandas==2.2.3
databricks-sdk==0.38.0
databricks-sql-connector==3.6.0
pydantic-settings==2.7.1
openpyxl==3.1.5
